join energy sums in one file
