--- generatorcore/industry2018/empty.py
@@ -1,8 +1,0 @@
-# pyright: strict
-from dataclasses import dataclass
-
-
-@dataclass
-class Empty:
-    # Used s_renew_hydrogen, s_renew_emethan
-    pass
